Improve JFA

+++ vplib/src/vox/naive.cu
@@ -30,10 +30,10 @@ __global__ void NaiveProcessing(const size_t numTriangles,
     std::pair<float, float> BB_X, BB_Y, BB_Z;
     CalculateBoundingBox(std::span<Position>(&facesVertices[0], 3), BB_X, BB_Y, BB_Z);
 
-    int startY = static_cast<int>(std::floor((BB_Y.first - grid.OriginY()) / grid.VoxelSize()));
-    int endY   = static_cast<int>(std::ceil((BB_Y.second - grid.OriginY()) / grid.VoxelSize()));
-    int startZ = static_cast<int>(std::floor((BB_Z.first - grid.OriginZ()) / grid.VoxelSize()));
-    int endZ   = static_cast<int>(std::ceil((BB_Z.second - grid.OriginZ()) / grid.VoxelSize()));
+    int startY = static_cast<int>(floorf((BB_Y.first - grid.OriginY()) / grid.VoxelSize()));
+    int endY   = static_cast<int>(ceilf((BB_Y.second - grid.OriginY()) / grid.VoxelSize()));
+    int startZ = static_cast<int>(floorf((BB_Z.first - grid.OriginZ()) / grid.VoxelSize()));
+    int endZ   = static_cast<int>(ceilf((BB_Z.second - grid.OriginZ()) / grid.VoxelSize()));
 
     Position edge0 = V1 - V0;
     Position edge1 = V2 - V0;
